Update student report details and styling
- Changed the report month/year format from short to long for better clarity.
- Updated the student report cover layout by adjusting grid column sizes and logo dimensions for improved visual balance.
- Modified the report title from "Term Exam" to "Annual Examination" to better reflect the report's purpose.
- Removed text transformation from the school title for a more consistent appearance.

+++ frontend/src/reportUtils.js
@@ -448,7 +448,7 @@ export const buildStudentReportData = ({
     displayDob: formatDateDisplay(student?.dateOfBirth),
     studentAge: formatAgeFromMonths(getAgeInMonths(student?.dateOfBirth)),
     averageAgeInClass: formatAgeFromMonths(averageAgeMonths),
-    reportMonthYear: new Date().toLocaleDateString(undefined, { month: 'short', year: 'numeric' }),
+    reportMonthYear: new Date().toLocaleDateString(undefined, { month: 'long', year: 'numeric' }),
     objectiveSections,
     marksheetRows,
     totalMax: totalMax > 0 ? totalMax : '',
